apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlmodel>=0.0.24",
    "alembic>=1.14.0,<1.15.0",
    "greenlet>=3.2.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    
]
